spiders works with weird import and alchemy errors
