--- popcorn_page/popcorn/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PopcornConfig(AppConfig):
-    name = 'popcorn'
