+++ popcorn_page/popcorn_page/settings.py
@@ -31,7 +31,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    'popcorn.apps.PopcornConfig',
+    'scrapy_app.apps.ScrapyAppConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
