+++ popcorn_page/scrapy_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ScrapyAppConfig(AppConfig):
+    name = 'scrapy_app'
