+++ popcorn_page/scrapy_app/management/commands/crawl_new_items.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from popcorn_page.popcorn.spiders.series import LostfilmSeriesSpider
+from popcorn.spiders.new_items import LostfilmNewSpider
 from scrapy.crawler import CrawlerProcess
 from scrapy.utils.project import get_project_settings
 
@@ -9,5 +9,5 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         process = CrawlerProcess(get_project_settings())
 
-        process.crawl(LostfilmSeriesSpider)
+        process.crawl(LostfilmNewSpider)
         process.start()
