+++ popcorn_page/scrapy_app/management/commands/crawl_series.py
@@ -1,0 +1,13 @@
+from django.core.management.base import BaseCommand
+from popcorn.spiders.series import LostfilmSeriesSpider
+from scrapy.crawler import CrawlerProcess
+from scrapy.utils.project import get_project_settings
+
+
+class Command(BaseCommand):
+
+    def handle(self, *args, **options):
+        process = CrawlerProcess(get_project_settings())
+
+        process.crawl(LostfilmSeriesSpider)
+        process.start()
